Refactor checklib.py as a module
Previously, comparelibs.py had to run a separate process for
checklib.py.  This meant loading a whole Python interpreter potentially
hundreds of times for a large library with a lot of changes, making the
script run fairly slowly.

This commit refactors checklib.py so it can be loaded as a module.
comparelibs.py is also changed to use checklib.py this way.  This change
makes comparelibs.py run in about 80% the time it used to (~33.5s vs
~27s) when comparing KiCad 4.0.7's device.lib to the current
kicad-symbols Device.lib.