Move rule descriptions to rule class docstrings
As I wrote in #220, the rule class docstrings seem like a more sensible
place for rule descriptions.  I offered to make this change a couple of
months ago, and have seen no objections, so here it is.

While moving the descriptions, I changed them all to match the one-line
rule names in KLC 3.0.11, the current revision.

In moving the descriptions out of the constructor, many rule classes no
longer need to define their own constructors at all, so overall this
reduces how much code we have to maintain.