Only conduct KLC-check on added symbols, not complete library
Currently if a commit adds a symbol, the KLC check is run for the whole
file, not only for the added symbol. Since the KLC check output is used
to mark merge requests as KLC compliant, this can be misleading if
errors were present before application of the merge request in questioni
in symbols that have not been touched by the merge request. As
comparelibs.py already existed and seemed to be intended for just that
purpose, this patch changes the call to the check to actually use the
results from comparelibs.py.