Only parse kicad_sym files of version 20211014
The symbol parser is not very robust when it tries to read older (v5) or
newer (v7) kicad symbol files.
To prevent weird python-stack-traces and make it clear to the user, in most
cases that will be contributors looking at pipeline output, we want a clean
error message.
To avoid adding code for all the known and yet unknown file format differences
just stop working at all if the version does not match.

Related: #368